Reorganized back-end.
