+++ Stock-Market-Sim/src/main/java/com/revature/models/Deprecated/GamePhase.java
@@ -1,3 +1,4 @@
+package com.revature.models.Deprecated;
 //package com.revature.models;
 //
 //import java.util.ArrayList;
