+++ Stock-Market-Sim/src/main/java/com/revature/models/Deprecated/loginTemplate.java
@@ -1,0 +1,5 @@
+package com.revature.models.Deprecated;
+
+public class loginTemplate {
+
+}
