--- Stock-Market-Sim/src/main/java/com/revature/models/loginTemplate.java
@@ -1,5 +1,0 @@
-package com.revature.models;
-
-public class loginTemplate {
-
-}
